Changed

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
 
-    public double calculate(double weightKg, double heightM) {
+    public int calculate(double weightKg, double heightM) {
         double index = weightKg / heightM / heightM;
         int result;
         result = (int) index;
