+++ src/Main.java
@@ -4,7 +4,7 @@ public class Main {
         BmiService service = new BmiService();
         double weightKg = 98; // вес в кг
         double heightM = 1.87; // рост в метрах
-        int bmi = (int) service.calculate(weightKg, heightM);
+        int bmi = service.calculate(weightKg, heightM);
         System.out.println("bmi-индекс: " + bmi);
     }
 }
